pyproject.toml
[project]
name = "tasklib"
version = "0.1.0"
description = "Simple, durable task queue for PostgreSQL"
readme = "README.md"
authors = [
    { name = "Jakub Kriz", email = "jakubykriz@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "sqlmodel>=0.0.18",
    "psycopg[binary]>=3.2",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
tasklib-worker = "tasklib.cli:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"
